Heat map - Scatter plot interaction implemented
Made two two-way sliders that correspond to the x-axis and y-axis of heat map. When the user moves the slider thumbs, the black rectangle will move above the heat map to show the range of x-axis and y-axis. Scatter plot will be re-rendered real time using the modified x-axis and y-axis by the slider. Moved the original ScatterPlot.jsx code to ScatterPlot_Legacy.jsx